requirements.txt
streamlit==1.15.0
requests==2.26.0
pillow==9.0.0
pandas==1.3.3
